cookiecutter-ml-template/{{cookiecutter.project_name}}/requirements.txt
mlflow
hydra-core
omegaconf
scikit-learn
